Move mmss to api/utils.go
